requirements.txt
numpy==1.26.2
Pillow==10.0.1
Pillow==10.1.0
streamlit==1.29.0
tensorflow==2.15.0
tensorflow_intel==2.15.0
